完善

+++ app/controller/FoodService/food.py
@@ -3,12 +3,13 @@ from flask import request
 from app.db import User, Food, db
 
 class foodservice():
+
     def food(self):
         page = int(request.args.get('page'))
         #SecretKey = '0a6b58441e5069288e0f95939a2c4375'
         #UserPhone = '2147483647'
         #page = 1
-        f = Food.query.paginate(page,10,False)
+        f = Food.query.query.filter_by(State = 1).paginate(page,10,False)
         p = f.items
         if len(p) > 0:
             array = {
@@ -35,6 +36,8 @@ class foodservice():
                 'msg': '信息为空',
                 'state': '0'
             }
+            return array
+
     def creat(self):
         form = request.form
         UserPhone = form.get('UserPhone')
@@ -44,37 +47,43 @@ class foodservice():
         if UserPhone and SecretKey:
             u = User.query.filter_by(UserPhone = UserPhone).first()
             if u.SecretKey == SecretKey:
-                FoodArea = form.get('FoodArea')
-                FoodInformation = form.args.get('FoodInformation')
-                FoodTime = form.args.get('FoodTime')
-                FoodWay = form.args.get('FoodWay')
-                #FoodArea = 'hhh'
-                #FoodInformation = 'hhh'
-                #FoodTime = '2016-07-28 15:31:41'
-                #FoodWay = 'hhh'
                 UserId = u.UserId
-                f =  Food()
-                f.UserId = UserId
-                f.FoodArea = FoodArea
-                f.FoodInformation = FoodInformation
-                f.FoodTime = FoodTime
-                f.FoodWay = FoodWay
-                f.State = 1
-                db.session.add(f)
-                db.session.commit()
-                state = '1'
-                msg = '创建成功'
+                myf = Food.query.filter_by(UserId = UserId, State = 1)
+                if myf is not None:
+                    msg = '已有约会'
+                    state = '0'
+                else:
+                    FoodArea = form.get('FoodArea')
+                    FoodInformation = form.args.get('FoodInformation')
+                    FoodTime = form.args.get('FoodTime')
+                    FoodWay = form.args.get('FoodWay')
+                    #FoodArea = 'hhh'
+                    #FoodInformation = 'hhh'
+                    #FoodTime = '2016-07-28 15:31:41'
+                    #FoodWay = 'hhh'
+                    f =  Food()
+                    f.UserId = UserId
+                    f.FoodArea = FoodArea
+                    f.FoodInformation = FoodInformation
+                    f.FoodTime = FoodTime
+                    f.FoodWay = FoodWay
+                    f.State = 1
+                    db.session.add(f)
+                    db.session.commit()
+                    state = '1'
+                    msg = '创建成功'
             else:
                 state = '0'
-                msg = '创建失败'
+                msg = '请登录'
         else:
             state = '0'
-            msg = '创建失败'
+            msg = '请登录'
         array = {
             'state' : state,
             'msg' : msg
         }
         return array
+
     def get(self):
         UserPhone = request.args.get('UserPhone')
         SecretKey = request.args.get('SecretKey')
@@ -88,7 +97,6 @@ class foodservice():
                 f = Food.query.filter_by(FoodId = FoodId).first()
                 if f.State == 1:
                     f.GetUser = u.UserId
-                    f.State = 0
                     msg = '成功'
                     state = '1'
                 else:
@@ -105,6 +113,7 @@ class foodservice():
             'state' : state
         }
         return array
+
     def cancle(self):
         UserPhone = request.args.get('UserPhone')
         SecretKey = request.args.get('SecretKey')
@@ -113,10 +122,18 @@ class foodservice():
             u = User.query.filter_by(UserPhone = UserPhone).first()
             if u.SecretKey == SecretKey:
                 f = Food.query.filter_by(FoodId = FoodId).first()
-                db.session.delete(f)
-                db.session.commit()
-                msg = '撤销成功'
-                state = '1'
+                if f.UserId == u.UserId:
+                    if f.State == 1:
+                        db.session.delete(f)
+                        db.session.commit()
+                        msg = '撤销成功'
+                        state = '1'
+                    else:
+                        msg = '当前不能撤销'
+                        state = '0'
+                else:
+                    msg = '非法操作'
+                    state = '0'
             else:
                 msg = '请登录'
                 state = '0'
@@ -128,6 +145,7 @@ class foodservice():
             'msg' : msg
         }
         return array
+
     def myfood(self):
         UserPhone = request.args.get('UserPhone')
         SecretKey = request.args.get('SecretKey')
@@ -166,6 +184,7 @@ class foodservice():
                 'state': '0'
             }
             return array
+
     def thisfood(self):
         FoodId = int(request.args.get('FoodId'))
         p = Food.query.filter_by(FoodId = FoodId).first()
